pyproject.toml
[project]
name = "sparse-index-tracking"
version = "0.1.0"
description = "End-to-End Learning for Sparse Index Tracking (CSCI 619)"
authors = [
    { name = "Aojie Yuan" },
    { name = "Haiyue Zhang" },
]
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "numpy>=1.26",
    "pandas>=2.1",
    "yfinance>=0.2.36",
    "cvxpy>=1.4",
    "cvxpylayers>=0.1.6",
    "scikit-learn>=1.4",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "wandb>=0.16",
    "pytest>=8.0",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "pre-commit"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
